simplify address API
Remove a bunch of abstractions that are not used yet.  AddressRegion can
be replaced by Bytes, AddressSpace can be moved directly in GuestMemory
instead of making it a marker trait.
